Added support for default_type on some constants. default_type is required in some EDSL context to know how to handle the constant without any other informations.

+++ modules/boost/simd/constant/include/boost/simd/toolbox/constant/constants/smallestposval.hpp
@@ -20,7 +20,7 @@ namespace boost { namespace simd
   {
     struct Smallestposval 
     { 
-      typedef int default_type;
+      typedef double default_type;
       template<class Target, class Dummy=void> 
       struct apply : meta::int_c<Target,1> {};  
     };
